ConnectionHelper moved one level up in eva.utils
